requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coc.py==2.3.1
discord.py==2.2.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
prisma==0.8.2
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
ujson==5.7.0
yarl==1.9.2
jishaku==2.5.1